+++ src/Modules/EcommerceGate.Core/Models/UserAddress.cs
@@ -5,11 +5,11 @@ namespace EcommerceGate.Core.Models
 {
     public class UserAddress : EntityBase
     {
-        public long UserId { get; set; }
+        public int UserId { get; set; }
 
         public virtual User User { get; set; }
 
-        public long AddressId { get; set; }
+        public int AddressId { get; set; }
 
         public virtual Address Address { get; set; }
